# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "McDee Platform API",
        "description": "Multifunctional Marketplace & Services API",
        "contact": {
            "email": "support@mcdee.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/vendor/mechanic/setup": {
            "post": {
                "tags": [
                    "Vendor - Mechanic"
                ],
                "summary": "Complete mechanic vendor registration",
                "description": "Complete Mechanic Registration",
                "operationId": "6cb42b4cc9582759d93dae37051d018d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "workshop_name": {
                                        "type": "string",
                                        "example": "FixIt Autos"
                                    },
                                    "services_offered": {
                                        "type": "string",
                                        "example": "Oil change, Engine diagnostics"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Ikeja, Lagos"
                                    },
                                    "contact_number": {
                                        "type": "string",
                                        "example": "+2348012345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mechanic profile completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mechanic profile completed successfully"
                                        },
                                        "mechanic": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "vendor_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "workshop_name": {
                                                    "type": "string",
                                                    "example": "FixIt Autos"
                                                },
                                                "services_offered": {
                                                    "type": "string",
                                                    "example": "Oil change, Engine diagnostics"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Ikeja, Lagos"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "+2348012345678"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/vendor/mechanic/profile": {
            "get": {
                "tags": [
                    "Vendor - Mechanic"
                ],
                "summary": "Get current mechanic profile",
                "description": "Get Mechanic Profile",
                "operationId": "b710d43655ffef1a8a78bfea1d7c64e9",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mechanic": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "vendor_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "workshop_name": {
                                                    "type": "string",
                                                    "example": "FixIt Autos"
                                                },
                                                "services_offered": {
                                                    "type": "string",
                                                    "example": "Oil change, Engine diagnostics"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Ikeja, Lagos"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "+2348012345678"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/vendor/register": {
            "post": {
                "tags": [
                    "Vendor"
                ],
                "summary": "Register vendor",
                "description": "Register a new vendor\n\nRegister either an individual or business vendor.",
                "operationId": "65afd508724830b605e16970e17b9808",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendor_type",
                                    "category"
                                ],
                                "properties": {
                                    "vendor_type": {
                                        "type": "string",
                                        "enum": [
                                            "individual",
                                            "business"
                                        ],
                                        "example": "individual"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "example": "McDee Motors"
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "rider",
                                            "mechanic",
                                            "apartment",
                                            "product_vendor",
                                            "service_vendor"
                                        ],
                                        "example": "rider"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendor registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vendor registered successfully"
                                        },
                                        "vendor": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "vendor_type": {
                                                    "type": "string",
                                                    "example": "individual"
                                                },
                                                "business_name": {
                                                    "type": "string",
                                                    "example": "McDee Motors"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "rider"
                                                },
                                                "is_verified": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only vendor-type users can register as vendors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}